feat: taskbar initial ts conversion
