+++ src/Taskbar/Taskbar.styles.tsx
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import * as styled from "styled-components";
 
 
 const MainContainer = styled.div`
